refactor: dropdown type 분리
