+++ src/components/dropdown/types/index.ts
@@ -1,0 +1,1 @@
+export type DropDownType = 'single' | 'multi';
